backend/multi-account/control-tower/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --no-emit-index-url multi-account/control-tower/requirements.in
#
attrs==25.3.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.242
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==45.2.0
    # via aws-cdk-lib
aws-cdk-lib==2.206.0
    # via -r multi-account/control-tower/requirements.in
cattrs==24.1.3
    # via jsii
constructs==10.4.2
    # via
    #   -r multi-account/control-tower/requirements.in
    #   aws-cdk-lib
importlib-resources==6.5.2
    # via jsii
jsii==1.112.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.14.1
    # via jsii
